carbonwise-backend/requirements.txt
fastapi 
uvicorn 
selenium 
pandas 
numpy 
pydantic
requests
python-dotenv